refactor: move the template fetch step out of lib.rs
Materializing a source into a temp dir and resolving which directory
inside it is the template is one step, and it was the last thing in
lib.rs carrying `#[cfg(feature = "git")]`. It moves to `src/fetch.rs`
with its tests; lib.rs is now free of feature-gated code and 490
lines shorter.
